feat: 채팅 popover 처리
