+++ components/chatting/ChattingPop.tsx
@@ -1,0 +1,14 @@
+import Chatting from './Chatting';
+
+import { PopoverContent } from '@/components/ui/popover';
+
+
+const ChattingPop = () => {
+	return (
+		<PopoverContent align='end' className='w-[25vw] min-w-[350px] py-0 z-0 text-black'>
+			<Chatting />
+		</PopoverContent>
+	);
+};
+
+export default ChattingPop;
